# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  message: string;
}

export interface ClassSummary {
  id: string;
  name: string;
  level: string;
  academicYear: string;
  pupilCount: number;
}

export interface Pupil {
  id: string;
  registrationNumber: string;
  firstName: string;
  lastName: string;
  classId: string;
  active: boolean;
}

export interface PupilImportRow {
  registrationNumber: string;
  firstName: string;
  lastName: string;
  dateOfBirth?: string | null;
  gender?: string | null;
}

export interface PupilImportRequest {
  /** @maxItems 1000 */
  rows: PupilImportRow[];
}

export interface PupilImportError {
  row: number;
  message: string;
}

export interface PupilImportResult {
  imported: number;
  skipped: number;
  errors: PupilImportError[];
}

export interface Assessment {
  id: string;
  title: string;
  date: string;
  classId: string;
  subject: string;
  level: string;
  competency: string;
  status: string;
}

export interface Objective {
  id: string;
  order: number;
  description: string;
}

export type AssessmentDetail = Assessment & ({
  support?: string | null;
  sessionObjectives?: string | null;
  objectives: Objective[];
});

export type EvaluationInputValue = typeof EvaluationInputValue[keyof typeof EvaluationInputValue];


export const EvaluationInputValue = {
  NotEvaluated: 'NotEvaluated',
  Acquired: 'Acquired',
  PartiallyAcquired: 'PartiallyAcquired',
  NotAcquired: 'NotAcquired',
} as const;

export interface EvaluationInput {
  pupilId: string;
  objectiveId: string;
  value: EvaluationInputValue;
}

export interface ObjectiveStatistics {
  objectiveId: string;
  evaluated: number;
  notEvaluated: number;
  acquired: number;
  partiallyAcquired: number;
  notAcquired: number;
  acquiredPercent: number;
}

export interface Dashboard {
  schoolName: string;
  academicYear: string;
  classes: ClassSummary[];
  recentAssessments: Assessment[];
}

/**
 * Resource not found
 */
export type NotFoundResponse = Error;

export type SaveBulkEvaluations200 = {
  saved: number;
};