# lib/api-zod/src/generated/types/pupilImportRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PupilImportRow } from './pupilImportRow';

export interface PupilImportRequest {
  /** @maxItems 1000 */
  rows: PupilImportRow[];
}