# lib/api-zod/src/generated/types/pupilImportResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PupilImportError } from './pupilImportError';

export interface PupilImportResult {
  imported: number;
  skipped: number;
  errors: PupilImportError[];
}